delete redundant match_repositoy.py
