+++ backend/app/repositories/match_repository.py
@@ -1,11 +1,13 @@
+from datetime import datetime
 from typing import List
 from uuid import UUID
 
 from sqlalchemy.future import select
 from sqlalchemy.orm import selectinload
+from sqlmodel import func
 
 from app.core.db import async_session
-from app.models import User
+from app.models import Article, User
 from app.models.match import Match
 
 
@@ -67,3 +69,22 @@ class MatchRepository:
             await session.commit()
             await session.refresh(match)
             return match
+
+    @staticmethod
+    async def get_recent_match_count_by_profile_id(
+        profile_id: UUID,
+        since: datetime,
+    ) -> int:
+        async with async_session() as session:
+            since = since.replace(tzinfo=None)
+            stmt = (
+                select(func.count().label("count"))
+                .select_from(Match)
+                .join(Article, Match.article_id == Article.id)
+                .where(
+                    Match.search_profile_id == profile_id,
+                    Article.published_at >= since,
+                )
+            )
+            result = await session.execute(stmt)
+            return result.scalar_one_or_none() or 0
